TRAIN-218: Implement generic Budget controller.

+++ src/main/java/za/co/neslotech/spring/trainspringbootinsight/model/Budget.java
@@ -18,7 +18,7 @@ public class Budget {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
     private Date created;
@@ -35,11 +35,11 @@ public class Budget {
     @ManyToOne
     private User user;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(final int id) {
+    public void setId(final Long id) {
         this.id = id;
     }
